Implement unified AppShell pattern with proper navbar integration
- Create UnifiedAppShell component with route-based configuration
- Extract MainNavigation and DashboardNavigation components
- Integrate StateSelector as proper AppShell navbar for forecast view
- Remove custom Layout component and sidebar implementation
- Update button styling to match navigation patterns
- Clean up unused code and imports
- Remove card-based state selector (no longer used)

Benefits:
- Consistent layout system across all routes
- Better performance with shared navigation state
- Proper AppShell navbar integration
- Unified styling and responsive behavior
