+++ app/src/components/dashboard/MyRespiLensDashboard.jsx
@@ -48,16 +48,17 @@ import {
   IconInfoCircle
 } from '@tabler/icons-react';
 
-const MyRespiLensDashboard = () => {
-  const [activeTab, setActiveTab] = useState('overview');
-  const [user] = useState({
+const MyRespiLensDashboard = ({ activeTab = 'overview', user = null }) => {
+  const defaultUser = {
     name: 'Dr. Sarah Johnson',
     email: 'sarah.johnson@health.state.gov',
     role: 'State Epidemiologist',
     organization: 'New York State Health Department',
     joinDate: '2023-08-15',
     avatar: null
-  });
+  };
+
+  const currentUser = user || defaultUser;
 
   const [dashboardData] = useState({
     favoriteLocations: ['US', 'NY', 'CA', 'FL'],
@@ -153,123 +154,19 @@ const MyRespiLensDashboard = () => {
   };
 
   return (
-    <AppShell
-      navbar={{ width: 250, breakpoint: 'sm' }}
-      header={{ height: 70 }}
-      padding="md"
-    >
-      <AppShell.Navbar p="md">
-        <Stack gap="xs">
-          <Button
-            variant={activeTab === 'overview' ? 'light' : 'subtle'}
-            leftSection={<IconDashboard size={16} />}
-            justify="start"
-            onClick={() => setActiveTab('overview')}
-          >
-            Overview
-          </Button>
-          <Button
-            variant={activeTab === 'activity' ? 'light' : 'subtle'}
-            leftSection={<IconActivity size={16} />}
-            justify="start"
-            onClick={() => setActiveTab('activity')}
-          >
-            Recent Activity
-          </Button>
-          <Button
-            variant={activeTab === 'forecastable' ? 'light' : 'subtle'}
-            leftSection={<IconTarget size={16} />}
-            justify="start"
-            onClick={() => setActiveTab('forecastable')}
-          >
-            Forecastable Stats
-          </Button>
-          <Button
-            variant={activeTab === 'bookmarks' ? 'light' : 'subtle'}
-            leftSection={<IconBookmark size={16} />}
-            justify="start"
-            onClick={() => setActiveTab('bookmarks')}
-          >
-            Bookmarks
-          </Button>
-          <Button
-            variant={activeTab === 'settings' ? 'light' : 'subtle'}
-            leftSection={<IconSettings size={16} />}
-            justify="start"
-            onClick={() => setActiveTab('settings')}
-          >
-            Settings
-          </Button>
-        </Stack>
-
-        <Divider my="md" />
-
-        {/* Quick Actions */}
-        <Stack gap="xs">
-          <Text size="sm" fw={500} c="dimmed">Quick Actions</Text>
-          <Button variant="light" size="xs" leftSection={<IconChartLine size={14} />}>
-            View Forecasts
-          </Button>
-          <Button variant="light" size="xs" leftSection={<IconBookmark size={14} />}>
-            Browse Narratives
-          </Button>
-          <Button variant="light" size="xs" leftSection={<IconTarget size={14} />}>
-            Play Forecastable
-          </Button>
-        </Stack>
-      </AppShell.Navbar>
-
-      <AppShell.Header p="md">
-        <Group justify="space-between" align="center">
-          <Group>
-            <ThemeIcon size="lg" variant="light">
-              <IconDashboard size={24} />
-            </ThemeIcon>
-            <Title order={3}>MyRespiLens</Title>
-          </Group>
-          
-          <Group gap="xs">
-            <ActionIcon variant="subtle" size="lg">
-              <IconBell size={20} />
-            </ActionIcon>
-            <Menu shadow="md" width={200}>
-              <Menu.Target>
-                <Group style={{ cursor: 'pointer' }}>
-                  <Avatar size="sm" radius="xl" />
-                  <Text size="sm" fw={500}>{user.name}</Text>
-                </Group>
-              </Menu.Target>
-              <Menu.Dropdown>
-                <Menu.Item leftSection={<IconUser size={14} />}>
-                  Profile
-                </Menu.Item>
-                <Menu.Item leftSection={<IconSettings size={14} />}>
-                  Settings
-                </Menu.Item>
-                <Menu.Divider />
-                <Menu.Item leftSection={<IconLogout size={14} />} color="red">
-                  Logout
-                </Menu.Item>
-              </Menu.Dropdown>
-            </Menu>
-          </Group>
-        </Group>
-      </AppShell.Header>
-
-      <AppShell.Main>
-        <Container size="xl">
+    <Container size="xl">
           {activeTab === 'overview' && (
             <Stack gap="md">
               {/* Welcome Section */}
               <Paper shadow="sm" p="lg" radius="md">
                 <Group justify="space-between" align="flex-start">
                   <div>
-                    <Title order={2} mb="xs">Welcome back, {user.name.split(' ')[1]}!</Title>
+                    <Title order={2} mb="xs">Welcome back, {currentUser.name.split(' ')[1]}!</Title>
                     <Text c="dimmed">
-                      {user.role} at {user.organization}
+                      {currentUser.role} at {currentUser.organization}
                     </Text>
                     <Text size="sm" c="dimmed" mt="xs">
-                      Member since {new Date(user.joinDate).toLocaleDateString()}
+                      Member since {new Date(currentUser.joinDate).toLocaleDateString()}
                     </Text>
                   </div>
                   <Badge variant="light" size="lg">
@@ -489,9 +386,7 @@ const MyRespiLensDashboard = () => {
               </Tabs>
             </Stack>
           )}
-        </Container>
-      </AppShell.Main>
-    </AppShell>
+    </Container>
   );
 };
